move exception class
