+++ src/java/com/palantir/cassandra/utils/MutationVerificationUtils.java
@@ -25,6 +25,7 @@ import java.util.Collection;
 import java.util.List;
 
 import com.google.common.annotations.VisibleForTesting;
+import org.apache.cassandra.exceptions.InvalidMutationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
